Minor refactoring and changing the compute shader to deal with objects that aren't hard coded
Refactored bestHit to closestHit and PixelOffset to PixelOffsetForAntiAliasing

Created a sphere struct to use instead of just float4.  Added a way to color different objects/add a texture to them using by adding a spectral index to the RayHit struct and a color index to the Sphere struct.

Set up code to allow for non-hard coded objects like the spheres[] array, and the game object Objects with the
RayTracingObjects script.